requirements.txt
flask==3.1.0
firebase-admin==6.6.0
python-dotenv==1.0.1
flask-wtf==1.2.2
email-validator==2.2.0
Werkzeug==3.1.3
Jinja2==3.1.5
itsdangerous==2.2.0
click==8.1.8
blinker==1.9.0
pyrebase4==4.8.0 